+++ packages/components-react/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.1.0-next.32](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.1.0-next.31...@esri/calcite-components-react@5.1.0-next.32) (2026-04-14)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.1.0-next.31](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.1.0-next.30...@esri/calcite-components-react@5.1.0-next.31) (2026-04-13)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [5.1.0-next.30](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.1.0-next.29...@esri/calcite-components-react@5.1.0-next.30) (2026-04-07)
 
 **Note:** Version bump only for package @esri/calcite-components-react